Address review feedback
- Move getConnectScreenErrorMessage to helpers/ alongside the other public
  connection helpers, re-exported from the action component for existing
  importers.
- Map jetpack_id and wpcom_bad_response so their raw messages (embedded
  response body / bare status number) never reach a fallback.
- My Jetpack onboarding form: fall back to the server-written error message
  (response.message, without the "(Status N)" suffix) before the generic
  sentence, preserving user-facing text like cannot_save_secrets and wpcom
  error descriptions.
- Add cross-reference comments between the helper and the Jetpack plugin's
  duplicate map in state-notices.jsx.
