+++ projects/js-packages/connection/index.jsx
@@ -37,7 +37,7 @@ export { default as thirdPartyCookiesFallbackHelper } from './helpers/third-part
 export { default as getCalypsoOrigin } from './helpers/get-calypso-origin';
 export * from './helpers/get-user-connection-url.ts';
 export { getReconnectErrorMessage } from './helpers/get-reconnect-error-message.ts';
-export { getConnectScreenErrorMessage } from './components/connect-screen/action';
+export { getConnectScreenErrorMessage } from './helpers/get-connect-screen-error-message';
 
 /**
  * Store
