+++ projects/plugins/jetpack/_inc/client/components/jetpack-notices/state-notices.jsx
@@ -25,6 +25,11 @@ class JetpackStateNotices extends Component {
 		this.setState( { showNotice: false } );
 	};
 
+	// Several codes below (e.g. `wpcom_408`, `wpcom_5??`, `wpcom_bad_response`,
+	// `register_http_request_failed`, `connection_disabled`) also have copy in the shared
+	// `getConnectScreenErrorMessage` helper in
+	// `projects/js-packages/connection/helpers/get-connect-screen-error-message.tsx`;
+	// when changing copy for a shared code, check that file too.
 	getErrorFromKey = key => {
 		const errorDesc = this.props.jetpackStateNoticesErrorDescription || false;
 		let message;
